Revert "fix: attack spam"
This reverts commit a61820a2fb1bac687aadcaee24e2c5255e56e21f.

+++ src/Attack.cpp
@@ -26,7 +26,7 @@ Attack::Attack(GameObject& associated,
     bool friendly, 
     float speed,
     float duration, 
-    float frameCount,
+    int frameCount,
     float frameTime
 ): Component(associated) {
     this->origin = origin;
